Moved code for nametags from cloudperms module to simplenametags plugin
