+++ cloudnet-examples/src/main/java/de/dytanic/cloudnet/examples/permissions/UpdateNameTagsExample.java
@@ -5,7 +5,7 @@ import de.dytanic.cloudnet.driver.event.EventListener;
 import de.dytanic.cloudnet.driver.event.EventPriority;
 import de.dytanic.cloudnet.driver.event.events.permission.PermissionUpdateUserEvent;
 import de.dytanic.cloudnet.driver.permission.IPermissionUser;
-import de.dytanic.cloudnet.ext.cloudperms.bukkit.BukkitCloudNetCloudPermissionsPlugin;
+import de.dytanic.cloudnet.ext.simplenametags.CloudNetSimpleNameTagsPlugin;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -21,7 +21,7 @@ public final class UpdateNameTagsExample {
 
     @EventHandler
     public void executeBukkitExampleOnPlayerJoinEvent(PlayerJoinEvent event) {
-        BukkitCloudNetCloudPermissionsPlugin.getInstance().updateNameTags(event.getPlayer());
+        CloudNetSimpleNameTagsPlugin.getInstance().updateNameTags(event.getPlayer());
         //Sets the nametags and don't overwrite the scoreboard rather the scoreboard will updated
     }
 
@@ -31,13 +31,13 @@ public final class UpdateNameTagsExample {
         Player player = Bukkit.getPlayer(event.getPermissionUser().getUniqueId());
 
         if (player != null) {
-            BukkitCloudNetCloudPermissionsPlugin.getInstance().updateNameTags(player);
+            CloudNetSimpleNameTagsPlugin.getInstance().updateNameTags(player);
         }
     }
 
     //For developers with a NickAPI or something like this
     public void nickExample(Player player) {
-        BukkitCloudNetCloudPermissionsPlugin.getInstance().updateNameTags(player, player1 -> {
+        CloudNetSimpleNameTagsPlugin.getInstance().updateNameTags(player, player1 -> {
             if (this.isNicked(player1)) {
                 return CloudNetDriver.getInstance().getPermissionManagement().getGroup("Default");
             }
