+++ cloudnet-modules/cloudnet-cloudperms/src/main/java/de/dytanic/cloudnet/ext/cloudperms/bukkit/listener/BukkitCloudNetCloudPermissionsPlayerListener.java
@@ -14,9 +14,11 @@ import org.bukkit.event.player.PlayerQuitEvent;
 
 public final class BukkitCloudNetCloudPermissionsPlayerListener implements Listener {
 
+    private final BukkitCloudNetCloudPermissionsPlugin plugin;
     private final IPermissionManagement permissionsManagement;
 
-    public BukkitCloudNetCloudPermissionsPlayerListener(IPermissionManagement permissionsManagement) {
+    public BukkitCloudNetCloudPermissionsPlayerListener(BukkitCloudNetCloudPermissionsPlugin plugin, IPermissionManagement permissionsManagement) {
+        this.plugin = plugin;
         this.permissionsManagement = permissionsManagement;
     }
 
@@ -28,7 +30,7 @@ public final class BukkitCloudNetCloudPermissionsPlayerListener implements Liste
                 event.setKickMessage(ChatColor.translateAlternateColorCodes('&', message));
             }, Bukkit.getOnlineMode());
 
-            BukkitCloudNetCloudPermissionsPlugin.getInstance().injectCloudPermissible(event.getPlayer());
+            plugin.injectCloudPermissible(event.getPlayer());
         }
     }
